Add files via upload
I used a Linear Regression model for prediction in this project. Linear Regression is a statistical method that aims to establish a linear relationship between input variables and a target variable. In this context, it was employed to make predictions based on the linear combination of features, providing valuable insights into numerical outcomes .